Add files via upload
RNA structures can be represented in three different levels of dimensions. If they are provided as a sequence, it will be the 1D representation.  1D representation of RNA lacks any interactions i.e. cWW, non-cWW, stacking, base-phosphate, base-ribose etc. When we go up the representation level by one in terms of the dimensionality, we will see 2D representation and this dimension is where the cWW interactions become apparent such as C-G and A-U. 

After a brief background info, the job of this program is to find interactions (they are called matchings) by taking into consideration one main thing: when we include C-G interaction, there is no need to include one more time as G-C. (Avoid repetitions) Thus, this program can use an (1D) RNA sequence as an input, and then it can produce the number of interactions in 2D RNA.